hypr: Move final autostarts to systemd, and fix exiting using PartOf
I copied the three lines from waybar service, but I believe it's the
PartOf that's doing the heavy lifting.

The symptom I was having was that after exiting Hyprland, the
graphical-session services weren't quitting due to
hyprland-session-shutdown.target, so they kept getting auto-restarted
(because of on-failure) *after* the logout, meaning they'd hit the
restart rate limit, and then logging back in would result in the
services not starting.

With PartOf, it should make those services stop while Hyprland quits,
and everything works cleanly.